added: tests for telemetry module, fixed import errors for spooky core binary
